Moved manual_search logic out of webserve into manual_search.py
* Fixed some issues regarding the manualSelectCheckCache().
* Fixed issue where cacheCheck was done while manual search results page was refreshed (timing issue).
* Moved some show queue status messages to show_queue.py.
